+++ tests/testthat/test-[.R
@@ -1,6 +1,6 @@
-subsette_tests <- system.file("tests", "testthat", "test-[.R", package = "subsette")
-if (subsette_tests == "") {
-    skip("Skip subsetting tests: subsette tests have to be installed")
+crochet_tests <- system.file("tests", "testthat", "test-[.R", package = "crochet")
+if (crochet_tests == "") {
+    skip("Skip subsetting tests: crochet tests have to be installed")
 } else {
-    source(subsette_tests, local = TRUE)
+    source(crochet_tests, local = TRUE)
 }